refactor notification code
